+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [6.1.4] - 2018-09-06
+
 ### Added
 
 - **Textarea** character countdown when maxLength is defined